requirements.txt
#selenium
pandas
#fastapi
scipy
#moviepy
#sqlalchemy
#pymysql
#django
google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib
python_dotenv~=1.2
